% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{ngs.cl}
\alias{ngs.cl}
\title{Assess feature selection of NGS}
\usage{
ngs.cl(
  dat.list,
  DF,
  method = c("randomForest", "svm"),
  pars = valipars(sampling = "cv", niter = 20, nreps = 5, strat = TRUE)
)
}
\arguments{
\item{dat.list}{a list consisting data set and class data}

\item{DF}{a setset of plot tittle for PCA and PLS}

\item{method}{classification methods. Only support \code{randomForest} and \code{svm}}

\item{pars}{sampling setting for classification}
}
\value{
a list with components of plots and classification results
}
\description{
Apply unsupervised and supervised lotting and classification to assess the
goodness of feture selection for count data.
}
