% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{ngs}
\alias{ngs}
\title{RNA-Seq statistical analysis}
\usage{
ngs(data, cls, com, method = "stats.TSPM", norm.method = "TMM")
}
\arguments{
\item{data}{a data frame with gene x replicate format}

\item{cls}{a vector for class/group information}

\item{com}{a character string for comparision}

\item{method}{statistical analysis methods for RNA-Seq}

\item{norm.method}{normalisation method}
}
\value{
a list of objects
}
\description{
Helper function for RNA-Seq statistical analysis
}
