% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{ngs.stats}
\alias{ngs.stats}
\title{Get group stats of normalised NGS data}
\usage{
ngs.stats(mat, grp, nf, method = "mean")
}
\arguments{
\item{mat}{raw count data with gene x replicate format}

\item{grp}{group information}

\item{nf}{normalisation factor}

\item{method}{data centre function}
}
\value{
a list including stats summary and normalised data set
}
\description{
Get group stats of normalised NGS data
}
