% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{norm.factor}
\alias{norm.factor}
\title{Calculate normalisation factor}
\usage{
norm.factor(
  data,
  method = c("DESeq", "TMM", "RLE", "UQ", "none"),
  norm.data = FALSE
)
}
\arguments{
\item{data}{a data frame with gene x replicate format}

\item{method}{differential analysis method}

\item{norm.data}{a logical indicating whether normalise data or not}
}
\value{
a list with normalisation factor and normalised data
}
\description{
Calculate normalisation factor
}
