% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{norm_factor}
\alias{norm_factor}
\title{Calculate normalisation factor}
\usage{
norm_factor(
  data,
  method = c("DESeq", "TMM", "RLE", "UQ", "none"),
  norm.data = FALSE
)
}
\arguments{
\item{data}{a data frame with gene x replicate format}

\item{method}{differential analysis method}

\item{norm.data}{a logical indicating whether normalise data or not}
}
\value{
a list with normalisation factor and normalised data if \code{norm.data}
is TRUE otherwise an vector of normalisation factor
}
\description{
Calculate normalisation factor
}
\examples{
## data set from R package 'pasilla' 
dn <- system.file("extdata/pasilla_gene_counts.tsv", package = "rnaseqdea")
data <- read.table(dn, header = TRUE, row.names = 1, quote = "", comment.char = "")
keep <- rowSums(data) > 6000
data <- data[keep, ]
res <- norm_factor(data, norm.data = TRUE)
res$nf
head(res$norm.data)

}
