% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{otu.wrapper}
\alias{otu.wrapper}
\title{Get OTUs table}
\usage{
otu.wrapper(res, pval = "padj", alpha = 0.1)
}
\arguments{
\item{res}{results from differential analysys}

\item{pval}{select which type of p-value to be used}

\item{alpha}{the threshold for p-value}
}
\value{
a list with some results
}
\description{
Get OTUs table
}
