% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{p.adj.f}
\alias{p.adj.f}
\title{Correct p-value with independent filtering}
\usage{
p.adj.f(pval, filter.stat, alpha = 0.1, method = "BH")
}
\arguments{
\item{pval}{an vector of p-calues}

\item{filter.stat}{A vector of stage-one filter statistics, or a function
which is able to compute this vector from data, if data is supplied. See
\code{\link[genefilter:filtered_p]{genefilter::filtered_p()}}.}

\item{alpha}{p-values threshold for selection}

\item{method}{p-calue correction method}
}
\value{
a list.
}
\description{
Correct p-value with independent filtering
}
