% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{p_adj_f}
\alias{p_adj_f}
\title{Correct p-value with independent filtering}
\usage{
p_adj_f(pval, filter.stat, alpha = 0.1, method = "BH")
}
\arguments{
\item{pval}{an vector of p-values}

\item{filter.stat}{A vector of stage-one filter statistics, or a function
which is able to compute this vector from data, if data is supplied. See
\code{\link[genefilter:filtered_p]{genefilter::filtered_p()}}.}

\item{alpha}{p-values threshold for selection}

\item{method}{p-value correction method}
}
\value{
a list.
}
\description{
Correct p-value with independent filtering
}
\details{
This function is modified from some code segments from
}
\examples{
## data set from R package 'pasilla' 
dn <- system.file("extdata/pasilla_gene_counts.tsv", package = "rnaseqdea")
data <- read.table(dn, header = TRUE, row.names = 1, quote = "", comment.char = "")
keep <- rowSums(data) > 6000
data <- data[keep, ]
cls <- c("untreated", "untreated", "untreated", "untreated", 
         "treated", "treated", "treated")
cls <- as.factor(cls)

## use limma's voom
res <- stats_voom(data, cls, com = levels(cls), norm.method = "TMM")
names(res)     # "stats", "rej.num", "data", "cls", "padjf"
res$rej.num 
head(res$stats)

## independent filtering for adjusted p-values
stats <- res$stats
pval <- stats$pval
names(pval) <- rownames(stats)
padj <- p_adj_f(pval, stats$mean, alpha = 0.1, method = "BH")
names(padj)
padj$rej

}
