% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Plot PCA of count data}
\usage{
plot_PCA(
  x,
  intgroup = "condition",
  ntop = 500,
  returnData = FALSE,
  select,
  sample
)
}
\arguments{
\item{x}{data set for plotting}

\item{intgroup}{interesting groups: a character vector of names in
colData(x) to use for grouping}

\item{ntop}{number of top genes to use for principal components, selected
by highest row variance}

\item{returnData}{should the function only return the data.frame of PC1
and PC2 with intgroup covariates for custom plotting (default is FALSE)}

\item{select, sample}{select which samplr for plotting}
}
\value{
a plot
}
\description{
Plot PCA of count datasizeFactors<-
}
