% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{rna_seq_cl}
\alias{rna_seq_cl}
\title{Assess feature selection of RNA-Seq data}
\usage{
rna_seq_cl(
  dat.list,
  DF,
  method = c("randomForest", "svm"),
  pars = valipars(sampling = "cv", niter = 20, nreps = 5, strat = TRUE),
  ep = 0
)
}
\arguments{
\item{dat.list}{a list consisting data set and class data}

\item{DF}{a setset of plot tittle for PCA and PLS}

\item{method}{classification methods. Only support \code{randomForest} and \code{svm}}

\item{pars}{sampling setting for classification}

\item{ep}{an integer for plotting ellipse. 1 and 2 for plotting overall
and group ellipse, respectively. Otherwise, none. This is for PCA and
PLS plot.}
}
\value{
a list with components of plots and classification results
}
\description{
Apply unsupervised and supervised lotting and classification to assess the
goodness of feture selection for count data.
}
