% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\docType{package}
\name{rnaseqdea-package}
\alias{rnaseqdea}
\alias{rnaseqdea-package}
\title{rnaseqdea: RNA-Seq Differential Expression Analysis (rnaseqdea)}
\description{
\code{rnaseqdea} provides some functions for RNA-Seq differential
expression analysis.
}
\section{Main functions}{

The main mtExtra provides more functions for metabolomics data analysis.
These functions are statistical analysis and plot methods with \code{ggplot2}
and \code{lattice}. It uses \code{tidyverse} as well as \code{reshape2} and \code{plyr}
packages.
}

\section{Package context}{

This package follows the principles of the "tidyverse" as mush as possible.
It also uses \code{melt} in \code{reshape2} if \code{tidyr} is complicated in some
circumstances.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wanchanglin/rnaseqdea}
  \item Report bugs at \url{https://github.com/wanchanglin/rnaseqdea/issues}
}

}
\author{
\strong{Maintainer}: Wanchang Lin \email{wanchanglin@hotmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
