% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{stats.edgeR}
\alias{stats.edgeR}
\alias{stats.DESeq2}
\alias{stats.NOISeq}
\alias{stats.NBPSeq}
\alias{stats.EBSeq}
\alias{stats.SAMseq}
\alias{stats.voom}
\alias{stats.test}
\alias{stats.TSPM}
\alias{stats}
\title{Wrapper function for RNA-Seq differential expression analysis}
\usage{
stats.edgeR(data, cls, com, norm.method = "TMM", ...)

stats.DESeq2(data, cls, com, norm.method = "TMM", ...)

stats.NOISeq(data, cls, com, norm.method = "TMM", ...)

stats.NBPSeq(data, cls, com, norm.method = "TMM", ...)

stats.EBSeq(data, cls, com, norm.method = "TMM", ...)

stats.SAMseq(data, cls, com, norm.method = "TMM", ...)

stats.voom(data, cls, com, norm.method = "TMM", ...)

stats.test(
  data,
  cls,
  com,
  norm.method = "TMM",
  test.method = "oneway.test",
  ...
)

stats.TSPM(data, cls, com, norm.method = "TMM", ...)
}
\arguments{
\item{data}{a data frame with gene x replicate format}

\item{cls}{a vector for class/group information}

\item{com}{a character string for comparision}

\item{norm.method}{normalisation method}

\item{...}{further parameters to be passed into modelling}

\item{test.method}{hypothesis testing method for \code{stats.test}, e.g.
\code{\link[=oneway.test]{oneway.test()}}, \code{\link[=kruskal.test]{kruskal.test()}},  \code{\link[=wilcox.test]{wilcox.test()}}, \code{\link[=t.test]{t.test()}}.}
}
\value{
a list with contents: \itemize{
\item stats statiststical summary
\item rej.num p-value rejection
\item data transposed and normalised data set
\item cls factor for class information
\item padjf p-values, adjusted p-values and adjusted p-values with
filtering
}
}
\description{
Call different methods with normalisation methods
}
\details{
\code{stats.edgeR} uses only \code{\link[edgeR:exactTest]{edgeR::exactTest()}} for modelling.

\code{stats.DESeq2} uses \code{\link[DESeq2:nbinomWaldTest]{DESeq2::nbinomWaldTest()}} only.
}
\examples{
library("pasilla")
dn <- system.file("extdata/pasilla_gene_counts.tsv", package = "pasilla")
data <- read.table(dn, header = TRUE, row.names = 1, quote = "", comment.char = "")
keep <- rowSums(data) > 6000
data <- data[keep, ]
cls <- c("untreated", "untreated", "untreated", "untreated", 
         "treated", "treated", "treated")
cls <- as.factor(cls)

## use DESeq2
res <- stats.DESeq2(data, cls, com = levels(cls), norm.method = "TMM")
names(res)     # "stats", "rej.num", "data", "cls", "padjf"
res$rej.num 
head(res$stats)

\dontrun{ 
## use SAMseq
set.seed(100) 
res <- stats.test(data, cls, com = levels(cls), norm.method = "RLE") 
res$rej.num 
head(res$stats)

## use hypothesis test
res <- stats.test(data, cls, com = levels(cls), norm.method = "UQ") 
## change test method
res <- stats.test(data, cls, com = levels(cls), norm.method = "UQ", 
                  test.method = wilcox.test)
res$rej.num
head(res$stats)
} 

}
