% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{vst.rlt.tr}
\alias{vst.rlt.tr}
\title{Transform count data}
\usage{
vst.rlt.tr(data, method = c("vst", "rlt"), ...)
}
\arguments{
\item{data}{a data frame with gene x replicate format}

\item{method}{two normalisation method \code{vst} and \code{rst}}

\item{...}{further parameters to \code{vst} and \code{rlt}}
}
\value{
normalised data
}
\description{
Transform count data
}
