% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{rna_seq_dea}
\alias{rna_seq_dea}
\title{RNA-Seq statistical analysis}
\usage{
rna_seq_dea(
  data,
  cls,
  com,
  stats.method = "stats_TSPM",
  norm.method = "TMM",
  ep = 0,
  ...
)
}
\arguments{
\item{data}{a data frame with gene x replicate format}

\item{cls}{a vector for class/group information}

\item{com}{a character string for comparision}

\item{stats.method}{statistical analysis methods for RNA-Seq}

\item{norm.method}{onormalisation method}

\item{ep}{an integer for plotting ellipse. 1 and 2 for plotting overall
and group ellipse, respectively. Otherwise, none. This is for PCA and
MDS plot.}

\item{...}{further parameters to \code{stats.method}}
}
\value{
a list with contents: \itemize{
\item ma.p MA plot
\item volcano.p ivolcano plot
\item hist.p histogram
\item data.s normalised data set with highly expressed variables
\item mad.p MAD plot
\item pca PCA values
\item pca.p PCA plot
\item box.p boxplot
}
}
\description{
Helper function for RNA-Seq statistical analysis
}
\examples{
## data set from R package 'pasilla' 
dn <- system.file("extdata/pasilla_gene_counts.tsv", package = "rnaseqdea")
data <- read.table(dn, header = TRUE, row.names = 1, quote = "", comment.char = "")
cls <- c("untreated", "untreated", "untreated", "untreated", 
         "treated", "treated", "treated")
cls <- as.factor(cls)

if (TRUE) {
  keep <- rowSums(data) > 6000
} else {
  ## Filter low expression tags
  ## cpms <- edgeR::cpm(data)
  cpms <- t(t(data) / (1e-6 * colSums(data)))
  keep <- rowSums(cpms > 0.5) >= 2            # from edgeR workfow
}
sum(keep)
data <- data[keep, ]

res <- rna_seq_dea(data, cls, com = levels(cls), stats.method = "stats_edgeR",
                   norm.method = "DESeq", ep = 2)
names(res)									 
head(res$stats)
res$rej.num
dim(res$data)
dim(res$data.s)

res$ma.p
res$volcano.p
res$hist.p
res$mds.p
res$pca.p
}
