% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{rna_seq_stats}
\alias{rna_seq_stats}
\title{Get group stats of normalised RNA-Seq data}
\usage{
rna_seq_stats(mat, grp, nf, method = "mean")
}
\arguments{
\item{mat}{raw count data with gene x replicate format}

\item{grp}{group information}

\item{nf}{normalisation factor}

\item{method}{data centre function}
}
\value{
a list including stats summary and normalised data set
}
\description{
Get group stats of normalised RNA-Seq data
}
\details{
The summary is based on the normalised data, excluding p-values
and their corrections. The p-values are computed by different modelling
methods such as \code{DEseq2} and \code{edgeR}.
}
\examples{
## data set from R package 'pasilla' 
dn <- system.file("extdata/pasilla_gene_counts.tsv", package = "rnaseqdea")
data <- read.table(dn, header = TRUE, row.names = 1, quote = "", comment.char = "")
keep <- rowSums(data) > 6000
data <- data[keep, ]
cls <- c("untreated", "untreated", "untreated", "untreated", 
         "treated", "treated", "treated")
cls <- as.factor(cls)

## get the normalisation factor
nf <- norm_factor(data)

## get summary on normalised data
res <- rna_seq_stats(data,grp = cls, nf = nf)
names(res)
head(res$stats)
}
