% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\docType{package}
\name{rnaseqdea-package}
\alias{rnaseqdea}
\alias{rnaseqdea-package}
\title{rnaseqdea: RNA-Seq Differential Expression Analysis (rnaseqdea)}
\description{
\code{rnaseqdea} provides functions for RNA-Seq differential expression analysis.
It implements wrapper functions to call differential statistical
models to perform analysis. This package also provides a unified
normalisation methods for all used modelling methods. Some plots are
implemented for these modelling.
}
\section{Wrapper functions for statistical analysis methods}{

\itemize{
\item DESeq2
\item edgeR
\item NBPSeq
\item EBSeq
\item NOISeq
\item SAMseq
\item voom+limma
\item TSPM
\item hypothesis testing
}
}

\section{Normalisation methods}{

\itemize{
\item Normal factor: \code{DESeq}, \code{TMM}, \code{RLE}, \code{UQ}. Use for modelling.
\item Variance stabilizing transformation (VST) and Regularized log
transformation (RLT) implemented in package \code{DESeq2}.  Use for the
classification and visualisation.
}
}

\section{Plots}{

\itemize{
\item MA plot
\item Volcano plot
\item Histogram of p-values
\item PCA plot
\item MDS plot
\item Boxplot
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wanchanglin/rnaseqdea}
  \item Report bugs at \url{https://github.com/wanchanglin/rnaseqdea/issues}
}

}
\keyword{internal}
