% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{samr.estimate.depth}
\alias{samr.estimate.depth}
\title{Estimate the sequencing depth}
\usage{
samr.estimate.depth(x, method = norm.method)
}
\arguments{
\item{x}{the original count matrix. p by n matrix of features, one
observation per column.}

\item{method}{normalisation method}
}
\value{
a normalisation factor
}
\description{
Estimate the sequencing depth of each experiment for sequencing data.
}
\details{
use this one to override the original
\code{\link[samr:samr.estimate.depth]{samr::samr.estimate.depth()}} for \code{\link[=stats.SAMseq]{stats.SAMseq()}}.
}
\examples{
## 'stats.SAMseq()' use this one to override original one:
assignInNamespace("samr.estimate.depth", samr.estimate.depth, ns = "samr")
}
