% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{stats.edgeR}
\alias{stats.edgeR}
\alias{stats.DESeq2}
\alias{stats.NOISeq}
\alias{stats.NBPSeq}
\alias{stats.EBSeq}
\alias{stats.SAMseq}
\alias{stats.voom}
\alias{stats.test}
\alias{stats.TSPM}
\alias{stats}
\title{Wrapper function for RNA-Seq differential expression analysis}
\usage{
stats.edgeR(data, cls, com, norm.method = "TMM", ...)

stats.DESeq2(data, cls, com, norm.method = "TMM", ...)

stats.NOISeq(data, cls, com, norm.method = "TMM", ...)

stats.NBPSeq(data, cls, com, norm.method = "TMM", ...)

stats.EBSeq(data, cls, com, norm.method = "TMM", ...)

stats.SAMseq(data, cls, com, norm.method = "TMM", ...)

stats.voom(data, cls, com, norm.method = "TMM", ...)

stats.test(
  data,
  cls,
  com,
  norm.method = "TMM",
  test.method = "oneway.test",
  ...
)

stats.TSPM(data, cls, com, norm.method = "TMM", ...)
}
\arguments{
\item{data}{a data frame with gene x replicate format}

\item{cls}{a vector for class/group information}

\item{com}{a character string for comparision}

\item{norm.method}{normalisation method}

\item{...}{further parameters to be passed into modelling}

\item{test.method}{hypothesis testing method for \code{stats.test}, e.g.
\code{\link[=oneway.test]{oneway.test()}}, \code{\link[=kruskal.test]{kruskal.test()}},  \code{\link[=wilcox.test]{wilcox.test()}}, \code{\link[=t.test]{t.test()}}.}
}
\value{
a list with contents: \itemize{
\item stats statiststical summary
\item rej.num p-value rejection
\item data transposed and normalised data set
\item cls factor for class information
\item padjf p-values, adjusted p-values and adjusted p-values with
filtering
}
}
\description{
Call different methods with normalisation methods
}
\details{
\code{stats.edgeR} uses only \code{\link[edgeR:exactTest]{edgeR::exactTest()}} for modelling.

\code{stats.DESeq2} uses \code{\link[DESeq2:nbinomWaldTest]{DESeq2::nbinomWaldTest()}} only.
}
