% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{stats_edgeR}
\alias{stats_edgeR}
\alias{stats_DESeq2}
\alias{stats_NOISeq}
\alias{stats_NBPSeq}
\alias{stats_EBSeq}
\alias{stats_SAMseq}
\alias{stats_voom}
\alias{stats_TSPM}
\alias{stats_test}
\alias{stats}
\title{Wrapper function for RNA-Seq differential expression analysis}
\usage{
stats_edgeR(data, cls, com, norm.method = "TMM", ...)

stats_DESeq2(data, cls, com, norm.method = "TMM", ...)

stats_NOISeq(data, cls, com, norm.method = "TMM", ...)

stats_NBPSeq(data, cls, com, norm.method = "TMM", ...)

stats_EBSeq(data, cls, com, norm.method = "TMM", ...)

stats_SAMseq(data, cls, com, norm.method = "TMM", ...)

stats_voom(data, cls, com, norm.method = "TMM", ...)

stats_TSPM(data, cls, com, norm.method = "TMM", ...)

stats_test(
  data,
  cls,
  com,
  norm.method = "TMM",
  test.method = "oneway.test",
  ...
)
}
\arguments{
\item{data}{a data frame with gene x replicate format}

\item{cls}{a vector for class/group information}

\item{com}{a character string for comparision}

\item{norm.method}{normalisation method}

\item{...}{further parameters to be passed into modelling}

\item{test.method}{hypothesis testing method for \code{stats_test}, e.g.
\code{\link[=oneway.test]{oneway.test()}}, \code{\link[=kruskal.test]{kruskal.test()}},  \code{\link[=wilcox.test]{wilcox.test()}}, \code{\link[=t.test]{t.test()}}.}
}
\value{
a list with contents: \itemize{
\item stats statiststical summary
\item rej.num p-value rejection
\item data transposed and normalised data set
\item cls factor for class information
\item padjf p-values, adjusted p-values and adjusted p-values with
filtering
}
}
\description{
Call different methods with normalisation methods
}
\details{
\code{stats_edgeR} uses only \code{\link[edgeR:exactTest]{edgeR::exactTest()}} for modelling.

\code{stats_DESeq2} uses \code{\link[DESeq2:nbinomWaldTest]{DESeq2::nbinomWaldTest()}} only.
}
\examples{
## data set from R package 'pasilla' 
dn <- system.file("extdata/pasilla_gene_counts.tsv", package = "rnaseqdea")
data <- read.table(dn, header = TRUE, row.names = 1, quote = "", comment.char = "")
keep <- rowSums(data) > 6000
data <- data[keep, ]
cls <- c("untreated", "untreated", "untreated", "untreated", 
         "treated", "treated", "treated")
cls <- as.factor(cls)

## use edgeR
res <- stats_edgeR(data, cls, com = levels(cls), norm.method = "TMM")
names(res)     # "stats", "rej.num", "data", "cls", "padjf"
res$rej.num 
head(res$stats)

\dontrun{ 
## use DESeq2
res <- stats_DESeq2(data, cls, com = levels(cls), norm.method = "DESeq")

## use SAMseq
set.seed(100) 
res <- stats_test(data, cls, com = levels(cls), norm.method = "RLE") 

## use hypothesis test
res <- stats_test(data, cls, com = levels(cls), norm.method = "UQ") 

## change test method
res <- stats_test(data, cls, com = levels(cls), norm.method = "UQ", 
                  test.method = wilcox.test)
} 

}
\references{
Paul L. Auer and Rebecca W Doerge, A Two-Stage Poisson Model for Testing
RNA-Seq Data, Statistical Applications in Genetics and Molecular Biology,
2011 (TSPM R codes: https://www.stat.purdue.edu/~doerge/software/TSPM.R)
}
