% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{stats.edgeR}
\alias{stats.edgeR}
\alias{stats.DESeq2}
\alias{stats.NOISeq}
\alias{stats.NBPSeq}
\alias{stats.EBSeq}
\alias{stats.SAMseq}
\alias{stats.voom}
\alias{stats.Wilcox}
\alias{stats.TSPM}
\alias{stats}
\title{Wrapper function for RNA-Seq differential expression analysis}
\usage{
stats.edgeR(data, cls, com, norm.method = "TMM")

stats.DESeq2(data, cls, com, norm.method = "TMM")

stats.NOISeq(data, cls, com, norm.method = "TMM")

stats.NBPSeq(data, cls, com, norm.method = "TMM")

stats.EBSeq(data, cls, com, norm.method = "TMM")

stats.SAMseq(data, cls, com, norm.method = "TMM")

stats.voom(data, cls, com, norm.method = "TMM")

stats.Wilcox(data, cls, com, norm.method = "TMM")

stats.TSPM(data, cls, com, norm.method = "TMM")
}
\arguments{
\item{data}{a data frame with gene x replicate format}

\item{cls}{a vector for class/group information}

\item{com}{a character string for comparision}

\item{norm.method}{normalisation method}
}
\value{
a list with:
\itemize{
\item istatistical summary
\item p-value rejection
\item data set
\item class i nformation
\item adjusted p-values
}
}
\description{
Call different methods with normalisation methods
}
\examples{
## To use [stats.SAMseq()], you have to override/replace function:
assignInNamespace("samr.estimate.depth", samr.estimate.depth, ns = "samr")
}
