% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqdea.R
\name{vst_rlt_tr}
\alias{vst_rlt_tr}
\title{Transform count data}
\usage{
vst_rlt_tr(data, method = c("vst", "rlt"), ...)
}
\arguments{
\item{data}{a data frame with gene x replicate format}

\item{method}{two normalisation method \code{vst} and \code{rst}}

\item{...}{further parameters to \code{vst} and \code{rlt}}
}
\value{
normalised data
}
\description{
Transform count data. The transformaed data can be used in visulaisation
and classification.
}
\examples{
library("pasilla")
dn <- system.file("extdata/pasilla_gene_counts.tsv", package = "pasilla")
data <- read.table(dn, header = TRUE, row.names = 1, quote = "", comment.char = "")
res <- vst_rlt_tr(data)  
}
